requirements.txt
alembic==1.2.1
aniso8601==8.0.0
asn1crypto==1.1.0
attrs==19.2.0
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.1.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
mysql-connector==2.2.9
numpy==1.17.3
pandas==0.25.2
passlib==1.7.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
PyMySQL==0.9.3
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
tqdm==4.40.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
